services/admin-service/requirements.txt
Flask
Flask-JWT-Extended
Flask-Cors 
requests
python-dotenv
gunicorn